Rename StorageAddonProvider to ExternalStorageService
