+++ app/serializers/external-storage-service.ts
@@ -1,10 +1,10 @@
 import OsfSerializer from './osf-serializer';
 
-export default class StorageAddonProviderSerializer extends OsfSerializer {
+export default class ExternalStorageServiceSerializer extends OsfSerializer {
 }
 
 declare module 'ember-data/types/registries/serializer' {
     export default interface SerializerRegistry {
-        'storage-addon-provider': StorageAddonProviderSerializer;
+        'external-storage-service': ExternalStorageServiceSerializer;
     } // eslint-disable-line semi
 }
