+++ mirage/serializers/external-storage-service.ts
@@ -1,0 +1,6 @@
+import ExternalStorageServiceModel from 'ember-osf-web/models/external-storage-service';
+
+import ApplicationSerializer from './application';
+
+export default class ConfiguredStorageAddonSerializer extends ApplicationSerializer<ExternalStorageServiceModel> {
+}
