--- mirage/serializers/storage-addon-provider.ts
@@ -1,6 +1,0 @@
-import StorageAddonProviderModel from 'ember-osf-web/models/storage-addon-provider';
-
-import ApplicationSerializer from './application';
-
-export default class ConfiguredStorageAddonSerializer extends ApplicationSerializer<StorageAddonProviderModel> {
-}
